# backend/api/openapi.yaml
openapi: 3.0.0
info:
  title: DevDish API
  version: 1.0.0
  description: A comprehensive recipe management API built with Node.js, Express, and MongoDB
  contact:
    name: DevDish API Support
    email: support@devdish.com
  license:
    name: MIT
    url: https://spdx.org/licenses/MIT.html
servers:
  - url: http://localhost:5000/api
    description: Development server
  - url: https://your-deployed-app.herokuapp.com/api
    description: Production server
...
paths:
...
  /recipes:
    get:
      tags:
        - Recipes
      summary: Get all recipes for current user
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 12
          description: Number of recipes per page
        - in: query
          name: category
          schema:
            type: string
            enum: [Breakfast, Lunch, Dinner, Dessert, Snack, Beverage, Other, All]
          description: Filter by category
        - in: query
          name: search
          schema:
            type: string
          description: Search in title, description, ingredients, or instructions
        - in: query
          name: difficulty
          schema:
            type: string
            enum: [Easy, Medium, Hard, All]
          description: Filter by difficulty
        - in: query
          name: maxTime
          schema:
            type: integer
          description: Maximum total cooking time in minutes
        - in: query
          name: tags
          schema:
            type: array
            items:
              type: string
          description: Filter by tags
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [createdAt, title, difficulty, totalTime]
            default: createdAt
          description: Field to sort by
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
      responses:
        '200':
          description: Recipes retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          recipes:
                            type: array
                            items:
                              $ref: '#/components/schemas/Recipe'
                          totalPages:
                            type: integer
                          currentPage:
                            type: integer
                          total:
                            type: integer
                          availableTags:
                            type: array
                            items:
                              type: string
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      tags:
        - Recipes
      summary: Create a new recipe
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeInput'
      responses:
        '201':
          description: Recipe created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          recipe:
                            $ref: '#/components/schemas/Recipe'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'
...
  /public/recipes:
    get:
      tags:
        - Public
      summary: Get all public recipes
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
          description: Page number
        - in: query
          name: limit
          schema:
            type: integer
            default: 12
          description: Number of recipes per page
        - in: query
          name: category
          schema:
            type: string
            enum: [Breakfast, Lunch, Dinner, Dessert, Snack, Beverage, Other, All]
          description: Filter by category
        - in: query
          name: search
          schema:
            type: string
          description: Search in title or description
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [createdAt, title, views]
            default: createdAt
          description: Field to sort by
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [asc, desc]
            default: desc
          description: Sort order
      responses:
        '200':
          description: Public recipes retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Success'
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          recipes:
                            type: array
                            items:
                              $ref: '#/components/schemas/Recipe'
                          totalPages:
                            type: integer
                          currentPage:
                            type: integer
                          total:
                            type: integer
        '500':
          $ref: '#/components/responses/ServerError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Recipe:
      type: object
      properties:
        _id:
          type: string
          description: Recipe ID
        title:
          type: string
          description: Recipe title
          maxLength: 100
        description:
          type: string
          description: Recipe description
          maxLength: 500
        ingredients:
          type: array
          items:
            type: object
            required:
              - name
              - quantity
              - unit
            properties:
              name:
                type: string
                description: Ingredient name
              quantity:
                type: number
                description: Ingredient quantity
                minimum: 0
              unit:
                type: string
                enum: [g, kg, ml, l, cup, tbsp, tsp, piece, pinch, to taste]
                description: Ingredient unit
              notes:
                type: string
                description: Additional notes about ingredient
        instructions:
          type: array
          items:
            type: string
            description: Cooking instruction step
        prepTime:
          type: number
          description: Preparation time in minutes
          minimum: 0
        cookTime:
          type: number
          description: Cooking time in minutes
          minimum: 0
        servings:
          type: number
          description: Number of servings
          minimum: 1
        difficulty:
          type: string
          enum: [Easy, Medium, Hard]
          description: Recipe difficulty level
        category:
          type: string
          enum: [Breakfast, Lunch, Dinner, Dessert, Snack, Beverage, Other]
          description: Recipe category
        tags:
          type: array
          items:
            type: string
            description: Recipe tags
        image:
          type: string
          description: Recipe image URL
        isPublic:
          type: boolean
          description: Whether the recipe is publicly visible
        publicId:
          type: string
          description: Public ID for sharing
        views:
          type: number
          description: Number of views
          minimum: 0
        likes:
          type: array
          items:
            type: string
          description: User IDs who liked the recipe
        userId:
          type: string
          description: User ID who created the recipe
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    RecipeInput:
      type: object
      required:
        - title
        - ingredients
        - instructions
        - prepTime
        - cookTime
        - servings
        - difficulty
        - category
      properties:
        title:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 500
        ingredients:
          type: array
          items:
            type: object
            required:
              - name
              - quantity
              - unit
            properties:
              name:
                type: string
              quantity:
                type: number
                minimum: 0
              unit:
                type: string
                enum: [g, kg, ml, l, cup, tbsp, tsp, piece, pinch, to taste]
              notes:
                type: string
        instructions:
          type: array
          items:
            type: string
        prepTime:
          type: number
          minimum: 0
        cookTime:
          type: number
          minimum: 0
        servings:
          type: number
          minimum: 1
        difficulty:
          type: string
          enum: [Easy, Medium, Hard]
        category:
          type: string
          enum: [Breakfast, Lunch, Dinner, Dessert, Snack, Beverage, Other]
        tags:
          type: array
          items:
            type: string
        image:
          type: string
        isPublic:
          type: boolean
          default: false
    Success:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          description: Success message
        data:
          type: object
          description: Response data
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Error message
        error:
          type: string
          description: Detailed error information
        errors:
          type: array
          items:
            type: string
          description: Validation errors
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: Not authorized to access this route
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: Validation failed
            errors:
              - Username must be at least 3 characters long
...
    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            success: false
            message: Internal server error